Create game() and adjust playRound()
I created the game() which will start a new game of rock, paper,
scissors. At the start of the game the value for playerLife and
computerLife should be set to 5. The while loop should now run
playRounds(), determine a winner or tie after each run and subtract
1 of the life value of the loser. This should be repeated until the
playerLife or computerLife reaches zero, which would end the game.

Since the values for playerSelection and computerSelection should be
declared new each round, I moved the declaration inside the playRound().

For the game() to track the result of each playRound(), I changed the
function return to a more useful variable.