Change play mode to button based
First I moved  the lifePoint variable declaration to outside the game()
so that they are more accessible and added some for the button.
I also added a EventListener to the buttons, that run the game() by
feeding it the weapon choice as an argument.
I further adjusted the point counting function and the determination
of the end of the game.
I added the resetGame() to reset for a new start of the game.